Move blkiodev package to types.
